[MASTERDIET-25] user achievement list added

+++ src/main/java/pl/agh/edu/master_diet/service/converter/RecentProductConverter.java
@@ -15,7 +15,7 @@ public final class RecentProductConverter
         return BaseProductInfo.builder()
                 .id(product.getId())
                 .name(product.getName())
-                .calories(product.getCalories().longValue())
+                .calories(product.getCalories())
                 .defaultValue(product.getDefaultValue().longValue())
                 .unit(product.getUnit())
                 .build();
